Refactor DumpsysAppOps
