+++ mvt/android/parsers/__init__.py
@@ -4,7 +4,6 @@
 #   https://license.mvt.re/1.1/
 
 from .dumpsys import (
-    parse_dumpsys_appops,
     parse_dumpsys_battery_daily,
     parse_dumpsys_battery_history,
     parse_dumpsys_receiver_resolver_table,
